@@ -1,16 +1,17 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <LangVersion>7.1</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="2.2.6" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.2.6" />
-	<PackageReference Include="Microsoft.EntityFrameworkCore.Abstractions" Version="2.2.6" />
-	<PackageReference Include="Microsoft.EntityFrameworkCore.Analyzers" Version="2.2.6" />
-	<PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.0-preview2.19525.5" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.0-preview2.19525.5" />
+	<PackageReference Include="Microsoft.EntityFrameworkCore.Abstractions" Version="3.1.0-preview2.19525.5" />
+	<PackageReference Include="Microsoft.EntityFrameworkCore.Analyzers" Version="3.1.0-preview2.19525.5" />
+	<PackageReference Include="Newtonsoft.Json" Version="12.0.3-beta2" />
+	<PackageReference Include="System.Data.SqlClient" Version="4.7.0" />
   </ItemGroup>
 
   <ItemGroup>
