@@ -5,9 +5,9 @@ namespace RavenNest.BusinessLogic.Game
 {
     public interface IAuthManager
     {
-        Task<AuthToken> AuthenticateAsync(string username, string password);
+        AuthToken Authenticate(string username, string password);
         AuthToken Get(string authToken);
-        Task SignUpAsync(string userId, string userLogin, string userDisplayName, string userEmail, string password);
-        Task<bool> CheckIfIsAdminAsync(AuthToken authToken);
+        void SignUp(string userId, string userLogin, string userDisplayName, string userEmail, string password);
+        bool IsAdmin(AuthToken authToken);
     }
 }
